Moved pkg/conditions to pkg/util/conditions
